Skilt ut beregning til egen pakke under behandling. Laget service og controller (med gamle navn for å ikke brekke frontend)
