+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/vedtak/Vedtak.kt
@@ -32,7 +32,6 @@ data class Vedtak(
         @Column(name = "aktiv", nullable = false)
         var aktiv: Boolean = true,
 
-
         @Column(name = "fk_forrige_vedtak_id")
         var forrigeVedtakId: Long? = null,
